Updated scripts
An error in the script (firmbackbone.py) resulted in many duplicates in the silver data. This then resulted in the same duplicates in the gold data. The issue is solved in the firmbackbone.py script and two record counts are added to the 01-import script (after reading the source file and after writing the silver data). Typically, the silver data should contain equal or less rows than the initial source data.